+++ README.md
@@ -1,2 +1,2 @@
 # SuperSlimeSoccer_JavaEdition
-Super Slime Soccer game, made with Java and implementing Simple Graphics Library with Keyboard Events
+Super Slime Soccer match, made with Java and implementing Simple Graphics Library with Keyboard Events
