--- src/io/codeforall/javatars_filhosdamain/GameKeyboardHandler.java
@@ -1,6 +1,0 @@
-package io.codeforall.javatars_filhosdamain;
-
-import org.academiadecodigo.simplegraphics.keyboard.KeyboardHandler;
-
-public interface GameKeyboardHandler extends KeyboardHandler {
-}
