--- src/io/codeforall/javatars_filhosdamain/MenuKeyboardHandler.java
@@ -1,7 +1,0 @@
-package io.codeforall.javatars_filhosdamain;
-
-import org.academiadecodigo.simplegraphics.keyboard.KeyboardHandler;
-
-public interface MenuKeyboardHandler extends KeyboardHandler {
-
-}
